-[DATAMANAGEMENT] Upgrading space location implementations to a better solution using only interface hierarchy.

+++ NeoCom.DataManagement/src/main/java/org/dimensinfin/eveonline/neocom/domain/space/Station.java
@@ -1,11 +1,7 @@
 package org.dimensinfin.eveonline.neocom.domain.space;
 
-import org.dimensinfin.eveonline.neocom.esiswagger.model.GetUniverseStationsStationIdOk;
-
 public interface Station extends SpaceSystem {
-	Integer getStationId();
-
-//	GetUniverseStationsStationIdOk getStation();
+	Long getStationId();
 
 	String getStationName();
 }
